Refactor topic_create.php to update CSS file path and move styles to a new location; remove unused topic_create.css and load_more_messages.php
